assessment_tools/requirements.txt
streamlit
pdf2image
Pillow
google-generativeai
google-genai
chromadb
sentence-transformers
openpyxl
python-dotenv
